+++ crates/cargo-zng/src/res/tool.rs
@@ -299,9 +299,12 @@ pub struct Tools {
     about: About,
 }
 impl Tools {
-    pub fn capture(local: &Path, cache: PathBuf, about: About) -> anyhow::Result<Self> {
+    pub fn capture(local: &Path, cache: PathBuf, about: About, verbose: bool) -> anyhow::Result<Self> {
         let mut tools = vec![];
         visit_tools(local, |t| {
+            if verbose {
+                println!("found tool `{}` in `{}`", t.name, t.path.display())
+            }
             tools.push(t);
             Ok(ControlFlow::Continue(()))
         })?;